services/infra-gateway/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
httpx==0.25.1
aio-pika==9.3.1
PyJWT==2.8.0
prometheus-client==0.17.1
structlog==23.2.0
python-multipart==0.0.6
